refactor(ingest): move google-workspace adapter to packages/plugins/
Moves GoogleWorkspaceAdapter from engram-core/src/ingest/adapters/ to
packages/plugins/google-workspace/, following the in-repo plugin model
established by the gerrit port (ADR-006 / DECISIONS.md).

engram-core's dependency surface no longer includes google-auth-library.
The CLI imports the adapter from @engram/plugin-google-workspace.
