+++ packages/plugins/google-workspace/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@engram/plugin-google-workspace",
+  "version": "0.1.0",
+  "description": "Google Workspace enrichment adapter for engram (in-repo plugin)",
+  "type": "module",
+  "main": "src/index.ts",
+  "scripts": {
+    "build": "echo 'plugin: no build step required'",
+    "test": "bun test"
+  },
+  "dependencies": {
+    "engram-core": "workspace:*",
+    "google-auth-library": "^9.0.0",
+    "ulid": "^2.3.0"
+  }
+}
